requirements.txt
# Oscar extensions
https://github.com/django-oscar/django-oscar-stores/archive/630aa146066cd327085fe9206dc50f6e520cc023.zip#egg=django-oscar-stores==1.0-dev
django-oscar-paypal==0.9.1
django-oscar-datacash==0.8.3

# We need PostGIS as the stores extension uses GeoDjango.
psycopg2==2.5.1


# To use Solr, we need pysolr
pysolr==3.2

django-extensions==1.6.1
raven==4.0.3
